LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-same-url-expected.txt


FAIL <a> to identical url is a replace navigation assert_equals: expected "replace" but got "push"
